requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.24.0
polars>=0.19.0
pyarrow>=10.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
prophet>=1.1.4
statsmodels>=0.14.0

# Hyperparameter Optimization
optuna>=3.4.0

# Experiment Tracking
mlflow>=2.7.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0

# Performance Optimization
psutil>=5.9.0
holidays>=0.34